live: move some utility functions to a separate module
